+++ plugins/EnhancedHotSwapEnabler/src/gradle/kotlin/IntelliJDcevmGradlePlugin.kt
@@ -24,10 +24,14 @@ import org.gradle.api.tasks.JavaExec
 import org.gradle.kotlin.dsl.withType
 
 
+@Suppress("unused")
 internal class IntelliJDcevmGradlePlugin : Plugin<Gradle> {
     override fun apply(target: Gradle) = target.allprojects {
         val enableDcevm = providers.environmentVariable(ENABLE_DCEVM_ENV_KEY).map { it.toBoolean() }
         val enableHotswapAgent = providers.environmentVariable(ENABLE_HOTSWAP_AGENT_ENV_KEY).map { it.toBoolean() }
+        // see org.jetbrains.plugins.gradle.service.task.GradleTaskManagerExtensionDebuggerBridge.Companion
+        // use environment variable to avoid doFirst ordering issues
+        val debuggerEnabled = providers.environmentVariable("DEBUGGER_ENABLED").map { it.toBoolean() }
         tasks.withType<JavaExec>().configureEach {
 
             val dcevmSupportProvider = javaLauncher.map { launcher ->
@@ -41,13 +45,13 @@ internal class IntelliJDcevmGradlePlugin : Plugin<Gradle> {
 
 
             doFirst {
-                if (args.none { it.startsWith("-agentlib:jdwp") }) return@doFirst
-                if (enableDcevm.get()) {
-                    val support = dcevmSupportProvider.get()
-                    if (support is DCEVMSupport.NeedsArgs)
-                        jvmArgs(support.args)
-                    logger.lifecycle("DCEVM enabled for task $name")
-                }
+                val taskNames = target.startParameter.taskNames
+                if (path !in taskNames && name !in taskNames) return@doFirst
+                if (!debuggerEnabled.get() || !enableDcevm.get()) return@doFirst
+                val support = dcevmSupportProvider.get()
+                if (support is DCEVMSupport.NeedsArgs)
+                    jvmArgs(support.args)
+                logger.lifecycle("DCEVM enabled for task $name")
                 if (enableHotswapAgent.get()) {
                     jvmArgs(JVM_OPTION_EXTERNAL_HOTSWAP_AGENT)
                     logger.lifecycle("HotswapAgent enabled for task $name")