+++ plugins/EnhancedHotSwapEnabler/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Configurations and settings supports
+
 ## [1.0.0] - 2025-08-16
 
 ### Added